split up homepage into components
