+++ src/app/components/Providers.tsx
@@ -1,5 +1,4 @@
 "use client";
-
 import { ReactNode } from "react";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { RainbowKitProvider } from "@rainbow-me/rainbowkit";
